Rename the SIO default reader / writer
